requirements.txt
Django==4.2.7
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt
psycopg2==2.9.3
django-environ==0.11.2
pillow==10.0.1
django-filter==24.3
django-cors-headers==3.13.0
django-pgtrigger==4.5.3
django-q2==1.7.2
django-storages==1.14.2
django-logic==0.1.3
boto3==1.28.69
factory-boy==3.3.0
Faker==19.2.0
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
celery==5.3.6
django-celery-beat==2.5.0
drf-spectacular==0.26.4
nested-multipart-parser==1.5.0
pyyaml==6.0.1
uritemplate==4.1.1
twisted>=23.10.0rc1 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
pytest-django==4.8.0
numpy==1.26.4
django-tree-queries==0.19.0
rules==3.4
gunicorn==23.0.0
